+++ SHE_PPT/python/SHE_PPT/table_formats/she_lensmc_chains.py
@@ -69,22 +69,22 @@ class SheLensMcChainsMeta(SheTableMeta):
 
     def __init__(self):
         # Store the less-used comments in a dict
-        super().__init__(comments = OrderedDict(((self.fits_version, None),
-                                                 (self.fits_def, None),
-                                                 (self.fits_version, None),
-                                                 (self.fits_def, None),
-                                                 (self.she_flag_version, None),
-                                                 (self.model_hash, None),
-                                                 (self.model_seed, None),
-                                                 (self.noise_seed, None),
-                                                 (self.observation_id, None),
-                                                 (self.pointing_id, "List of pointing IDs"),
-                                                 (self.observation_time, None),
-                                                 (self.tile_id, None),
-                                                 (self.method, "Shear estimation method used to generate these chains"),
-                                                 (self.len_chain, None),
-                                                 (self.valid, "0: Not tested; 1: Pass; -1: Fail")
-                                                 )))
+        super().__init__(comments=OrderedDict(((self.fits_version, None),
+                                               (self.fits_def, None),
+                                               (self.fits_version, None),
+                                               (self.fits_def, None),
+                                               (self.she_flag_version, None),
+                                               (self.model_hash, None),
+                                               (self.model_seed, None),
+                                               (self.noise_seed, None),
+                                               (self.observation_id, None),
+                                               (self.pointing_id, "List of pointing IDs"),
+                                               (self.observation_time, None),
+                                               (self.tile_id, None),
+                                               (self.method, "Shear estimation method used to generate these chains"),
+                                               (self.len_chain, None),
+                                               (self.valid, "0: Not tested; 1: Pass; -1: Fail")
+                                               )))
 
 
 class SheLensMcChainsFormat(SheTableFormat):
@@ -100,66 +100,66 @@ class SheLensMcChainsFormat(SheTableFormat):
         # Table column labels and properties
 
         self.ID = self.set_column_properties(
-            "OBJECT_ID", dtype = ">i8", fits_dtype = "K")
+            "OBJECT_ID", dtype=">i8", fits_dtype="K")
 
         self.fit_flags = self.set_column_properties(
-            "SHE_LENSMC_FIT_FLAGS", dtype = ">i8", fits_dtype = "K")
+            "SHE_LENSMC_FIT_FLAGS", dtype=">i8", fits_dtype="K")
         self.val_flags = self.set_column_properties(
-            "SHE_LENSMC_VAL_FLAGS", dtype = ">i8", fits_dtype = "K")
+            "SHE_LENSMC_VAL_FLAGS", dtype=">i8", fits_dtype="K")
         self.fit_class = self.set_column_properties(
-            "SHE_LENSMC_FIT_CLASS", dtype = ">i2", fits_dtype = "I")
+            "SHE_LENSMC_FIT_CLASS", dtype=">i2", fits_dtype="I")
         self.weight = self.set_column_properties(
-            "SHE_LENSMC_CHAINS_SHEAR_WEIGHT", dtype = ">f4", fits_dtype = "E")
+            "SHE_LENSMC_CHAINS_SHEAR_WEIGHT", dtype=">f4", fits_dtype="E")
         self.shape_weight = self.set_column_properties(
-            "SHE_LENSMC_CHAINS_SHAPE_WEIGHT", dtype = ">f4",
-            fits_dtype = "E", is_optional = True)
+            "SHE_LENSMC_CHAINS_SHAPE_WEIGHT", dtype=">f4",
+            fits_dtype="E", is_optional=True)
         self.e_var = self.set_column_properties(
-            "SHE_LENSMC_E_VAR", dtype = ">f4", fits_dtype = "E", is_optional = True)
+            "SHE_LENSMC_E_VAR", dtype=">f4", fits_dtype="E", is_optional=True)
         self.shape_noise = self.set_column_properties(
-            "SHE_LENSMC_ASSUMED_SHAPE_NOISE", dtype = ">f4",
-            fits_dtype = "E", is_optional = True)
+            "SHE_LENSMC_ASSUMED_SHAPE_NOISE", dtype=">f4",
+            fits_dtype="E", is_optional=True)
 
         self.g1 = self.set_column_properties(
-            "SHE_LENSMC_G1_CHAIN", dtype = ">f4", fits_dtype = "E",
-            length = total_chain_length)
+            "SHE_LENSMC_G1_CHAIN", dtype=">f4", fits_dtype="E",
+            length=total_chain_length)
         self.g2 = self.set_column_properties(
-            "SHE_LENSMC_G2_CHAIN", dtype = ">f4", fits_dtype = "E",
-            length = total_chain_length)
+            "SHE_LENSMC_G2_CHAIN", dtype=">f4", fits_dtype="E",
+            length=total_chain_length)
 
         self.ra = self.set_column_properties(
-            "SHE_LENSMC_UPDATED_RA_CHAIN", is_optional = True, comment = "deg",
-            fits_dtype = "D", length = total_chain_length)
+            "SHE_LENSMC_UPDATED_RA_CHAIN", is_optional=True, comment="deg",
+            fits_dtype="D", length=total_chain_length)
         self.dec = self.set_column_properties(
-            "SHE_LENSMC_UPDATED_DEC_CHAIN", is_optional = True, comment = "deg",
-            fits_dtype = "D", length = total_chain_length)
+            "SHE_LENSMC_UPDATED_DEC_CHAIN", is_optional=True, comment="deg",
+            fits_dtype="D", length=total_chain_length)
 
         # lensmc specific columns
         self.re = self.set_column_properties(
-            "SHE_LENSMC_RE_CHAIN", is_optional = True, dtype = ">f4", fits_dtype = "E",
-            length = total_chain_length)
+            "SHE_LENSMC_RE_CHAIN", is_optional=True, dtype=">f4", fits_dtype="E",
+            length=total_chain_length)
         self.flux = self.set_column_properties(
-            "SHE_LENSMC_FLUX_CHAIN", is_optional = True, dtype = ">f4", fits_dtype = "E",
-            length = total_chain_length)
+            "SHE_LENSMC_FLUX_CHAIN", is_optional=True, dtype=">f4", fits_dtype="E",
+            length=total_chain_length)
         self.magnitude = self.set_column_properties(
-            "SHE_LENSMC_MAGNITUDE_CHAIN", is_optional = True, dtype = ">f4", fits_dtype = "E",
-            length = total_chain_length)
+            "SHE_LENSMC_MAGNITUDE_CHAIN", is_optional=True, dtype=">f4", fits_dtype="E",
+            length=total_chain_length)
         self.bulge_frac = self.set_column_properties(
-            "SHE_LENSMC_BULGE_FRAC_CHAIN", is_optional = True, dtype = ">f4",
-            fits_dtype = "E", length = total_chain_length)
+            "SHE_LENSMC_BULGE_FRAC_CHAIN", is_optional=True, dtype=">f4",
+            fits_dtype="E", length=total_chain_length)
         self.snr = self.set_column_properties(
-            "SHE_LENSMC_SNR_CHAIN", is_optional = True, dtype = ">f4", fits_dtype = "E",
-            length = total_chain_length)
+            "SHE_LENSMC_SNR_CHAIN", is_optional=True, dtype=">f4", fits_dtype="E",
+            length=total_chain_length)
         self.lr = self.set_column_properties(
-            "SHE_LENSMC_LR_CHAIN", is_optional = True, dtype = ">f4", fits_dtype = "E",
-            length = total_chain_length)
+            "SHE_LENSMC_LR_CHAIN", is_optional=True, dtype=">f4", fits_dtype="E",
+            length=total_chain_length)
         self.chi2 = self.set_column_properties(
-            "SHE_LENSMC_CHI2", is_optional = True, dtype = ">f4", fits_dtype = "E")
+            "SHE_LENSMC_CHI2", is_optional=True, dtype=">f4", fits_dtype="E")
         self.dof = self.set_column_properties(
-            "SHE_LENSMC_DOF", is_optional = True, dtype = ">f4", fits_dtype = "E")
+            "SHE_LENSMC_DOF", is_optional=True, dtype=">f4", fits_dtype="E")
         self.acc = self.set_column_properties(
-            "SHE_LENSMC_ACCEPTANCE", is_optional = True, dtype = ">f4", fits_dtype = "E")
+            "SHE_LENSMC_ACCEPTANCE", is_optional=True, dtype=">f4", fits_dtype="E")
         self.nexp = self.set_column_properties(
-            "SHE_LENSMC_NEXP", is_optional = True, dtype = ">f4", fits_dtype = "E")
+            "SHE_LENSMC_NEXP", is_optional=True, dtype=">f4", fits_dtype="E")
 
         self._finalize_init()
 
@@ -262,7 +262,7 @@ def initialise_lensmc_chains_table(mer_final_catalog = None,
     """
 
     assert (mer_final_catalog is None) or (
-        is_in_format(mer_final_catalog, mfc_tf, strict = False))
+        is_in_format(mer_final_catalog, mfc_tf, strict=False))
 
     if optional_columns is None:
         optional_columns = []
@@ -272,16 +272,16 @@ def initialise_lensmc_chains_table(mer_final_catalog = None,
             if colname not in tf.all:
                 raise ValueError("Invalid optional column name: " + colname)
 
-    lensmc_chains_table = init_table(tf, optional_columns = optional_columns, init_cols = init_cols, size = size)
+    lensmc_chains_table = init_table(tf, optional_columns=optional_columns, init_cols=init_cols, size=size)
 
-    lensmc_chains_table.meta = make_lensmc_chains_table_header(model_hash = model_hash,
-                                                               model_seed = model_seed,
-                                                               noise_seed = noise_seed,
-                                                               observation_id = observation_id,
-                                                               pointing_id = pointing_id,
-                                                               observation_time = observation_time,
-                                                               method = method,
-                                                               tile_id = tile_id, )
+    lensmc_chains_table.meta = make_lensmc_chains_table_header(model_hash=model_hash,
+                                                               model_seed=model_seed,
+                                                               noise_seed=noise_seed,
+                                                               observation_id=observation_id,
+                                                               pointing_id=pointing_id,
+                                                               observation_time=observation_time,
+                                                               method=method,
+                                                               tile_id=tile_id, )
 
     assert is_in_format(lensmc_chains_table, tf)
